Changed volume calculation to just gridding, calculations will need to be in a third stage
